Expose openairs.json file publicly
This will allow reading the file from the deployed site instead of from
the local filesystem while rendering pages on the server at build time.
This in turn should allow to take advantage of the fetch API's Next.js
extensions for time-based revalidation, which is necessary to only show
recent/upcoming events on the home page.